+++ multimodal/tests/unittests/others_2/test_problem_types.py
@@ -26,10 +26,11 @@ from autogluon.multimodal.problem_types import PROBLEM_TYPES_REG
 )
 def test_get_problem_type(name):
     problem_prop = PROBLEM_TYPES_REG.get(name)
-    print(problem_prop)
     assert problem_prop.name == PROBLEM_TYPES_REG.get(problem_prop.name).name
 
 
 @pytest.mark.parametrize("name", PROBLEM_TYPES_REG.list_keys())
-def fetch_predictor_via_problem_types(name):
+def test_problem_type_in_predictor(name):
     predictor = MultiModalPredictor(problem_type=name)
+    assert predictor.problem_type == PROBLEM_TYPES_REG.get(name).name
+
